Clarify inconsistent specified semantics for array getters
Most array getters were specified to write out the number of elements
written if input capacity is nonzero, but cited the "Buffer Size
Parameters" section which specifies that the required capacity must be
written (and XR_ERROR_SIZE_INSUFFICIENT returned) if insufficient but
nonzero input capacity was specified. This clarifies the boilerplate
in each declaration to be consistent with the fundamentals section on
buffer sizes.

Co-Authored-By: Ryan Pavlik <ryan.pavlik@collabora.com>